api/pyproject.toml
[tool.poetry]
name = "ai-bridge"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "0.3.23"
office365-rest-python-client = "2.6.2"
html2text = "2020.1.16"
requests = "2.31.0"
pypdf = "3.17.4"
aiofiles = "23.2.1"
atlassian-python-api = "3.41.9"
pymongo = "4.12.0"
python-dotenv = "1.1.0"
simple-salesforce = "1.12.5"
pydantic = "2.9.0"
cachetools = "5.3.3"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
opencensus-ext-azure = "1.1.13"
opencensus = "0.11.4"
openai = "1.58.1"
openpyxl = "3.1.5"
oracledb = "2.4.1"
oci = "2.135.2"
jsonref = "1.1.0"
azure-ai-inference = "1.0.0b6"
groq = "0.12.0"
litestar = "^2.16.0"
uvicorn = "0.32.0"
apscheduler = "^3.11.0"
aiodebug = "2.3.0"
aiocache = "0.12.3"
motor = "3.7.0"
langchain-community = "0.3.21"
qdrant-client = "1.13.3"
pandas = "2.2.3"
yake = "^0.4.8"
nest-asyncio = "^1.6.0"
opentelemetry-api = "1.35.0"
opentelemetry-sdk = "1.35.0"
opentelemetry-exporter-otlp = "1.35.0"
mcp = "1.9.2"
cryptography = "45.0.4"
advanced-alchemy = "^1.4.5"
structlog = "^25.4.0"
aiosqlite = "^0.21.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
pgvector = "^0.4.1"

[tool.poetry.group.dev.dependencies]
ruff = "0.9.10"
pytest = "7.4.2"
pytest-mock = "3.11.1"
pytest-cov = "4.1.0"
# mypy = "1.10.0"
# pylint = "2.17.7"
# black = "23.9.1"
# isort = "5.10.1"

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "ASYNC",  "LOG", "PIE", "E"]
ignore = ["E501", "D401", "B006", 'C416', 'C418', 'UP007', 'S101', 'I001']
# select = ["E", "F", "W", "B", "I", "C", "ASYNC", "UP", "S", "LOG", "PIE", "SLF", "SIM", "TC", "ARG", "N", "PERF", "E"]

[tool.black]
line-length = 150
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
    "tests"
]